moved the lib additional code from .vcxproj into hpp.
From a library creation point of view

The person who takes the code and uses it should clearly know where the lib is being used.

If you add it to the .vcxproj field, you cannot clearly see it without looking at the separate comment field.

My library is independent except for hpp and cpp of the same name.

In conclusion,
Since the library has a strong purpose of being used by porting it to other code,

I thought it would be good to add this in the code so that it can be clearly distinguished.