update encrypt namespace
AES encryption and decryption functions have been added using CryptoPP.

It's in the prototype stage right now, so don't touch the code, there are unnecessary casts, so we're thinking of some logic to improve it.

I didn't know how to use the Hex Decodor class, but with this commit, I learned how to use the Encodor and Decodor classes.

+
Additionally,
Fixed the problem of not loading library from hash.hpp file (add pragma comment)

Additionally,
Added default parameters in the capture function in the image namespace.

Please check the change history when finished.